auth/requirements.txt
alembic==1.12.1
blinker==1.6.3
click==8.1.7
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
SQLAlchemy==2.0.22
SQLAlchemy-Utils==0.41.1
typing_extensions==4.8.0
Werkzeug==3.0.0
